Frameworks: fcharge for all (1/2)
Thanks roman for the basic idea to use an overlay
to set the fast charge path so each device can set
the propper path to the custom kernel.
As well this commit pays attention if the user
flashed a kernel which does not support fastcharge
or is using another path. Option for tile in settings will be removed
as well.

not 100% happy with the whole fast charge implementation...but I am lazy :p
...whatever maybe I will recode that all on a later point. It
is working as intended like it is.

+++ packages/SystemUI/src/com/android/systemui/quicksettings/FChargeTile.java
@@ -41,9 +41,6 @@ public class FChargeTile extends QuickSettingsTile {
     public static FChargeTile mInstance;
     private final boolean DBG = false;
 
-    public static final String FAST_CHARGE_DIR = "/sys/kernel/fast_charge";
-    public static final String FAST_CHARGE_FILE = "force_fast_charge";
-
     protected boolean enabled = false;
 
     public static QuickSettingsTile getInstance(Context context, LayoutInflater inflater,
@@ -65,13 +62,19 @@ public class FChargeTile extends QuickSettingsTile {
             public void onClick(View v) {
                 try {
                         enabled = !isFastChargeOn();
-                        File fastcharge = new File(FAST_CHARGE_DIR, FAST_CHARGE_FILE);
-                        FileWriter fwriter = new FileWriter(fastcharge);
-                        BufferedWriter bwriter = new BufferedWriter(fwriter);
-                        bwriter.write(enabled ? "1" : "0");
-                        bwriter.close();
-                        Settings.System.putInt(mContext.getContentResolver(),
-                             Settings.System.FCHARGE_ENABLED, enabled ? 1 : 0);
+                        String fchargePath = mContext.getResources()
+                                .getString(com.android.internal.R.string.config_fastChargePath);
+                        if (!fchargePath.isEmpty()) {
+                            File fastcharge = new File(fchargePath);
+                            if (fastcharge.exists()) {
+                                FileWriter fwriter = new FileWriter(fastcharge);
+                                BufferedWriter bwriter = new BufferedWriter(fwriter);
+                                bwriter.write(enabled ? "1" : "0");
+                                bwriter.close();
+                                Settings.System.putInt(mContext.getContentResolver(),
+                                     Settings.System.FCHARGE_ENABLED, enabled ? 1 : 0);
+                            }
+                        }
                     } catch (IOException e) {
                         if (DBG) Log.e("FChargeToggle", "Couldn't write fast_charge file");
                         Settings.System.putInt(mContext.getContentResolver(),
@@ -93,20 +96,26 @@ public class FChargeTile extends QuickSettingsTile {
 
     public boolean isFastChargeOn() {
         try {
-            File fastcharge = new File(FAST_CHARGE_DIR, FAST_CHARGE_FILE);
-            FileReader reader = new FileReader(fastcharge);
-            BufferedReader breader = new BufferedReader(reader);
-            String line = breader.readLine();
-            breader.close();
-            Settings.System.putInt(mContext.getContentResolver(),
-                    Settings.System.FCHARGE_ENABLED, line.equals("1") ? 1 : 0);
-            return (line.equals("1"));
+            String fchargePath = mContext.getResources()
+                    .getString(com.android.internal.R.string.config_fastChargePath);
+            if (!fchargePath.isEmpty()) {
+                File fastcharge = new File(fchargePath);
+                if (fastcharge.exists()) {
+                    FileReader reader = new FileReader(fastcharge);
+                    BufferedReader breader = new BufferedReader(reader);
+                    String line = breader.readLine();
+                    breader.close();
+                    Settings.System.putInt(mContext.getContentResolver(),
+                            Settings.System.FCHARGE_ENABLED, line.equals("1") ? 1 : 0);
+                    return (line.equals("1"));
+                }
+            }
         } catch (IOException e) {
             if (DBG) Log.e("FChargeToggle", "Couldn't read fast_charge file");
             Settings.System.putInt(mContext.getContentResolver(),
                  Settings.System.FCHARGE_ENABLED, 0);
-            return false;
         }
+        return false;
     }
 
     private void updateTileState() {
