+++ packages/SystemUI/src/com/android/systemui/statusbar/powerwidget/FChargeButton.java
@@ -40,8 +40,6 @@ import android.os.PowerManager;
 
 public class FChargeButton extends PowerButton {
 
-    public static final String FAST_CHARGE_DIR = "/sys/kernel/fast_charge";
-    public static final String FAST_CHARGE_FILE = "force_fast_charge";
     protected boolean on = false;
 
     private static final List<Uri> OBSERVED_URIS = new ArrayList<Uri>();
@@ -67,13 +65,19 @@ public class FChargeButton extends PowerButton {
     protected void toggleState(Context context) {
     try {
             on = !isFastChargeOn();
-            File fastcharge = new File(FAST_CHARGE_DIR, FAST_CHARGE_FILE);
-            FileWriter fwriter = new FileWriter(fastcharge);
-            BufferedWriter bwriter = new BufferedWriter(fwriter);
-            bwriter.write(on ? "1" : "0");
-            bwriter.close();
-            Settings.System.putInt(mContext.getContentResolver(),
-                     Settings.System.FCHARGE_ENABLED, on ? 1 : 0);
+            String fchargePath = mContext.getResources()
+                    .getString(com.android.internal.R.string.config_fastChargePath);
+            if (!fchargePath.isEmpty()) {
+                File fastcharge = new File(fchargePath);
+                if (fastcharge.exists()) {
+                    FileWriter fwriter = new FileWriter(fastcharge);
+                    BufferedWriter bwriter = new BufferedWriter(fwriter);
+                    bwriter.write(on ? "1" : "0");
+                    bwriter.close();
+                    Settings.System.putInt(mContext.getContentResolver(),
+                             Settings.System.FCHARGE_ENABLED, on ? 1 : 0);
+                }
+            }
         } catch (IOException e) {
             Log.e("FChargeToggle", "Couldn't write fast_charge file");
             Settings.System.putInt(mContext.getContentResolver(),
@@ -94,18 +98,24 @@ public class FChargeButton extends PowerButton {
 
     public boolean isFastChargeOn() {
         try {
-            File fastcharge = new File(FAST_CHARGE_DIR, FAST_CHARGE_FILE);
-            FileReader reader = new FileReader(fastcharge);
-            BufferedReader breader = new BufferedReader(reader);
-            String line = breader.readLine();
-            breader.close();
-            return (line.equals("1"));
+            String fchargePath = mContext.getResources()
+                    .getString(com.android.internal.R.string.config_fastChargePath);
+            if (!fchargePath.isEmpty()) {
+                File fastcharge = new File(fchargePath);
+                if (fastcharge.exists()) {
+                    FileReader reader = new FileReader(fastcharge);
+                    BufferedReader breader = new BufferedReader(reader);
+                    String line = breader.readLine();
+                    breader.close();
+                    return (line.equals("1"));
+                }
+            }
         } catch (IOException e) {
             Log.e("FChargeToggle", "Couldn't read fast_charge file");
             Settings.System.putInt(mContext.getContentResolver(),
                  Settings.System.FCHARGE_ENABLED, 0);
-            return false;
         }
+        return false;
     }
 
     @Override
